requirements.txt
google-generativeai==0.6.0
Pillow==10.3.0
requests==2.32.3
